@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:get/get.dart';
 import 'package:photo_view/photo_view.dart';
+import 'package:punjab_tourism/utils.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 import 'package:youtube_player_iframe/youtube_player_iframe.dart';
@@ -17,10 +18,10 @@ class DashboardSliderWidget extends StatelessWidget {
   final List<Attributes> attributes;
   var selectedIndex = 2.obs;
   DashboardSliderWidget({
-    Key key,
-    this.attributes,
+    Key? key,
+    required this.attributes,
   }) : super(key: key);
-  YoutubePlayerController controller;
+  YoutubePlayerController? controller;
   var youTubeId = ''.obs;
   @override
   Widget build(BuildContext context) {
@@ -41,34 +42,33 @@ class DashboardSliderWidget extends StatelessWidget {
                       if (attributes.elementAt(selectedIndex.value).type == "V") {
                         final id = Helper.convertUrlToId(attributesData.link);
                         String thumb = "https://img.youtube.com/vi/$id/0.jpg";
+
                         controller = YoutubePlayerController(
-                          initialVideoId: id,
                           params: YoutubePlayerParams(
-                            autoPlay: true,
-                            startAt: Duration(seconds: 0),
+                            origin: id,
                             showControls: true,
                             showFullscreenButton: true,
                             loop: false,
                             strictRelatedVideos: false,
                             enableJavaScript: false,
                             playsInline: false,
-                            desktopMode: true,
                           ),
                         );
-                        controller.onEnterFullscreen = () {
-                          SystemChrome.setPreferredOrientations([
-                            DeviceOrientation.landscapeLeft,
-                            DeviceOrientation.landscapeRight,
-                          ]);
-                        };
-                        controller.onExitFullscreen = () {
-                          SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp]);
-                          Future.delayed(const Duration(seconds: 1), () {
-                            controller.play();
-                          });
-                          Future.delayed(const Duration(seconds: 5), () {
-                            SystemChrome.setPreferredOrientations(DeviceOrientation.values);
-                          });
+                        controller?.onFullscreenChange = (entered) {
+                          if (entered) {
+                            SystemChrome.setPreferredOrientations([
+                              DeviceOrientation.landscapeLeft,
+                              DeviceOrientation.landscapeRight,
+                            ]);
+                          } else {
+                            SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp]);
+                            Future.delayed(const Duration(seconds: 1), () {
+                              controller!.playVideo();
+                            });
+                            Future.delayed(const Duration(seconds: 5), () {
+                              SystemChrome.setPreferredOrientations(DeviceOrientation.values);
+                            });
+                          }
                         };
                         Get.to(
                           Scaffold(
@@ -80,19 +80,20 @@ class DashboardSliderWidget extends StatelessWidget {
                               elevation: 0,
                               automaticallyImplyLeading: true,
                               centerTitle: true,
-                              title: "360 Video".text.textStyle(Get.textTheme.headline1.copyWith(color: Get.theme.indicatorColor)).make(),
+                              title: "360 Video".text.textStyle(headline1().copyWith(color: Get.theme.indicatorColor)).make(),
                             ),
                             backgroundColor: Get.theme.primaryColor,
                             body: InkWell(
                               onTap: () async {
-                                youTubeId.value = id;
+                                youTubeId.value = id!;
                                 if (Platform.isIOS) {
                                   if (await canLaunchUrlString(attributesData.link)) {
                                     await launch(attributesData.link, forceSafariVC: false);
                                   }
                                 } else {
-                                  controller.play();
-                                  controller.hideTopMenu();
+                                  controller?.playVideo();
+                                  // todo Manish kumar - check this todo before release
+                                  //controller?.hideTopMenu();
                                 }
                               },
                               child: Container(
@@ -155,7 +156,7 @@ class DashboardSliderWidget extends StatelessWidget {
                               elevation: 0,
                               automaticallyImplyLeading: true,
                               centerTitle: true,
-                              title: "Picture".text.textStyle(Get.textTheme.headline1.copyWith(color: Get.theme.indicatorColor)).make(),
+                              title: "Picture".text.textStyle(headline1().copyWith(color: Get.theme.indicatorColor)).make(),
                             ),
                             backgroundColor: Get.theme.primaryColor,
                             body: Container(
@@ -215,7 +216,7 @@ class DashboardSliderWidget extends StatelessWidget {
                                   height: 300,
                                   child: attributesData.title.text
                                       .textStyle(
-                                        Get.textTheme.headline2.copyWith(
+                                        headline2().copyWith(
                                           color: Get.theme.highlightColor,
                                         ),
                                       )
@@ -245,7 +246,7 @@ class DashboardSliderWidget extends StatelessWidget {
                       quarterTurns: 3,
                       child: currentValue.title.text
                           .textStyle(
-                            Get.textTheme.headline3.copyWith(
+                            headline3().copyWith(
                               color: Get.theme.indicatorColor.withOpacity(selectedIndex.value == index ? 1 : 0.5),
                             ),
                           )
