@@ -3,12 +3,15 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:get/get.dart';
+import 'package:punjab_tourism/utils.dart';
 
 import '../core.dart';
 
 class LoginView extends GetView<AuthController> {
-  LoginView({Key key}) : super(key: key);
+  LoginView({Key? key}) : super(key: key);
   CommonService commonService = Get.find();
+  final GlobalKey<FormState> loginFormKey = GlobalKey<FormState>();
+
   @override
   Widget build(BuildContext context) {
     SystemChrome.setSystemUIOverlayStyle(
@@ -32,13 +35,13 @@ class LoginView extends GetView<AuthController> {
             ),
           ),
           child: Form(
-            key: controller.loginFormKey,
+            key: loginFormKey,
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 commonService.labelData.value.data.login.text
-                    .textStyle(Get.theme.textTheme.headline1.copyWith(
+                    .textStyle(headline1().copyWith(
                       color: Get.theme.indicatorColor,
                       fontSize: 35,
                     ))
@@ -50,12 +53,14 @@ class LoginView extends GetView<AuthController> {
                     Platform.isAndroid
                         ? InkWell(
                             onTap: () async {
-                              String hint = await controller.autoFillSMS.hint;
+                              String? hint = await controller.autoFillSMS.hint;
                               if (hint != null && hint != '') {
                                 controller.checkMobileNumber(hint);
                               }
                             },
                             child: Container(
+                              padding: EdgeInsets.only(
+                                  left: 16, right: 16, top: 12, bottom: 12),
                               decoration: BoxDecoration(
                                 borderRadius: BorderRadius.circular(100),
                                 gradient: LinearGradient(
@@ -68,15 +73,12 @@ class LoginView extends GetView<AuthController> {
                                     stops: const [0.0, 1.0],
                                     tileMode: TileMode.clamp),
                               ),
-                              child: commonService
-                                  .labelData.value.data.verifyMobileNo.text
-                                  .textStyle(
-                                    Get.textTheme.headline3.copyWith(
-                                      color: Get.theme.highlightColor,
-                                    ),
-                                  )
-                                  .make()
-                                  .pSymmetric(h: 25, v: 25),
+                              child: Text(
+                                commonService
+                                    .labelData.value.data.verifyMobileNo,
+                                style:
+                                    TextStyle(color: Get.theme.highlightColor),
+                              ),
                             ),
                           )
                             .pOnly(bottom: Platform.isAndroid ? 30 : 0)
@@ -93,8 +95,7 @@ class LoginView extends GetView<AuthController> {
                                 ),
                               ),
                               commonService.labelData.value.data.or.text
-                                  .textStyle(
-                                      Get.theme.textTheme.headline1.copyWith(
+                                  .textStyle(headline1().copyWith(
                                     color: Get.theme.indicatorColor,
                                   ))
                                   .make()
@@ -114,11 +115,11 @@ class LoginView extends GetView<AuthController> {
                           color: Get.theme.indicatorColor,
                           fontWeight: FontWeight.w300),
                       keyboardType: TextInputType.emailAddress,
-                      validator: (String input) {
-                        if (input.isEmpty) {
+                      validator: (String? input) {
+                        if (input != null && input.isEmpty) {
                           return commonService
                               .labelData.value.data.emailFieldValidation;
-                        } else if (!GetUtils.isEmail(input)) {
+                        } else if (!GetUtils.isEmail(input!)) {
                           return commonService
                               .labelData.value.data.emailNotValid;
                         } else {
@@ -142,7 +143,7 @@ class LoginView extends GetView<AuthController> {
                               size: 20, color: Get.theme.colorScheme.primary),
                         ),
                         errorStyle: const TextStyle(fontSize: 15),
-                        hintStyle: Get.theme.textTheme.bodyText1.copyWith(
+                        hintStyle: bodyText1.copyWith(
                             color: Get.theme.indicatorColor.withOpacity(0.3)),
                         border: UnderlineInputBorder(
                             borderRadius: BorderRadius.circular(0),
@@ -164,8 +165,8 @@ class LoginView extends GetView<AuthController> {
                                     Get.theme.indicatorColor.withOpacity(0.5))),
                         errorBorder: UnderlineInputBorder(
                             borderRadius: BorderRadius.circular(0),
-                            borderSide: BorderSide(
-                                width: 1, color: Get.theme.errorColor)),
+                            borderSide:
+                                BorderSide(width: 1, color: errorColor)),
                         focusedErrorBorder: UnderlineInputBorder(
                             borderRadius: BorderRadius.circular(0),
                             borderSide: BorderSide(
@@ -176,102 +177,43 @@ class LoginView extends GetView<AuthController> {
                     ).pOnly(bottom: 20),
                     InkWell(
                       onTap: () {
-                        if (controller.loginFormKey.currentState.validate()) {
+                        if (loginFormKey.currentState!.validate()) {
                           controller.sendOtpToEmail();
                         }
                       },
                       child: Container(
-                        decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(100),
-                          gradient: LinearGradient(
-                              colors: [
-                                Get.theme.colorScheme.primary,
-                                Get.theme.colorScheme.secondary,
-                              ],
-                              begin: const FractionalOffset(0.0, 0.0),
-                              end: const FractionalOffset(0.0, 1.0),
-                              stops: const [0.0, 1.0],
-                              tileMode: TileMode.clamp),
-                        ),
-                        child: commonService.labelData.value.data.sendOtp.text
-                            .textStyle(
-                              Get.textTheme.headline2.copyWith(
-                                color: Get.theme.highlightColor,
-                              ),
-                            )
-                            .make()
-                            .pSymmetric(h: 20, v: 15),
-                      ),
+                          padding: EdgeInsets.only(
+                              left: 16, right: 16, top: 12, bottom: 12),
+                          decoration: BoxDecoration(
+                            borderRadius: BorderRadius.circular(100),
+                            gradient: LinearGradient(
+                                colors: [
+                                  Get.theme.colorScheme.primary,
+                                  Get.theme.colorScheme.secondary,
+                                ],
+                                begin: const FractionalOffset(0.0, 0.0),
+                                end: const FractionalOffset(0.0, 1.0),
+                                stops: const [0.0, 1.0],
+                                tileMode: TileMode.clamp),
+                          ),
+                          child: Text(
+                            commonService.labelData.value.data.sendOtp,
+                            style: TextStyle(color: Get.theme.highlightColor),
+                          )),
                     ).objectCenterRight(),
                   ],
                 ).pOnly(bottom: 30),
-                Row(
-                  children: [
-                    Expanded(
-                      child: Container(
-                        width: Get.width,
-                        color: Get.theme.colorScheme.secondary,
-                        height: 2,
-                      ),
-                    ),
-                    commonService.labelData.value.data.or.text
-                        .textStyle(Get.theme.textTheme.headline1.copyWith(
-                          color: Get.theme.indicatorColor,
-                        ))
-                        .make()
-                        .pSymmetric(h: 10),
-                    Expanded(
-                      child: Container(
-                        width: Get.width,
-                        color: Get.theme.colorScheme.secondary,
-                        height: 2,
-                      ),
-                    ),
-                  ],
-                ).pOnly(bottom: 20),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: [
-                    InkWell(
-                      onTap: () => controller.loginWithGoogle(),
-                      child: Image.asset(
-                        "assets/images/g-logo.png",
-                        width: 35,
-                        height: 35,
-                      ),
-                    ),
-                    // InkWell(
-                    //   onTap: () => controller.loginWithFB(),
-                    //   child: Image.asset(
-                    //     "assets/images/f-logo.png",
-                    //     width: 35,
-                    //     height: 35,
-                    //   ),
-                    // ),
-                    if (Platform.isIOS)
-                      InkWell(
-                        onTap: () => controller.loginWithApple(),
-                        child: Image.asset(
-                          "assets/images/apple.png",
-                          width: 40,
-                          height: 40,
-                        ),
-                      )
-                  ],
-                ),
-                SizedBox(
-                  height: 30,
-                ),
                 Row(
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: [
                     InkWell(
                       onTap: () {
-                        Helper.launchURL("https://bhagatsm.dwgroup.in/policies");
+                        Helper.launchURL(
+                            "https://bhagatsm.dwgroup.in/policies");
                       },
                       child: "Privacy Policy"
                           .text
-                          .textStyle(Get.theme.textTheme.bodyMedium.copyWith(
+                          .textStyle(Get.theme.textTheme.bodyMedium!.copyWith(
                             color: Get.theme.indicatorColor,
                           ))
                           .make(),
@@ -287,7 +229,7 @@ class LoginView extends GetView<AuthController> {
                       },
                       child: "Terms & Conditions"
                           .text
-                          .textStyle(Get.theme.textTheme.bodyMedium.copyWith(
+                          .textStyle(Get.theme.textTheme.bodyMedium!.copyWith(
                             color: Get.theme.indicatorColor,
                           ))
                           .make(),
