@@ -1,19 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-// import 'package:flutter_blue_plus/flutter_blue_plus.dart';
-import 'package:flutter_svg/svg.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:get/get.dart';
 import 'package:pin_code_fields/pin_code_fields.dart';
+import 'package:punjab_tourism/utils.dart';
+
 import '../core.dart';
 
 class VerifyOTPView extends GetView<AuthController> {
-  VerifyOTPView({Key key}) : super(key: key);
+  VerifyOTPView({Key? key}) : super(key: key);
   CommonService commonService = Get.find();
   @override
   Widget build(BuildContext context) {
     SystemChrome.setSystemUIOverlayStyle(
-      SystemUiOverlayStyle(statusBarColor: Get.theme.primaryColor, statusBarIconBrightness: Brightness.dark),
+      SystemUiOverlayStyle(
+          statusBarColor: Get.theme.primaryColor,
+          statusBarIconBrightness: Brightness.dark),
     );
     return Scaffold(
       backgroundColor: Get.theme.primaryColorDark,
@@ -40,14 +42,14 @@ class VerifyOTPView extends GetView<AuthController> {
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             commonService.labelData.value.data.verifyOtp.text
-                .textStyle(Get.theme.textTheme.headline1.copyWith(
+                .textStyle(headline1().copyWith(
                   color: Get.theme.indicatorColor,
                   fontSize: 35,
                 ))
                 .make()
                 .pOnly(bottom: 10),
             commonService.labelData.value.data.verifyOtpDesc.text
-                .textStyle(Get.theme.textTheme.bodyText1.copyWith(
+                .textStyle(bodyText1.copyWith(
                   fontSize: 17,
                   color: Get.theme.indicatorColor.withOpacity(0.5),
                 ))
@@ -61,9 +63,10 @@ class VerifyOTPView extends GetView<AuthController> {
               enablePinAutofill: true,
               enabled: true,
               keyboardType: TextInputType.number,
+              textStyle: TextStyle(color: Colors.white),
               pinTheme: PinTheme(
                 inactiveColor: Get.theme.colorScheme.primary,
-                errorBorderColor: Get.theme.errorColor,
+                errorBorderColor: errorColor,
                 shape: PinCodeFieldShape.box,
                 borderRadius: BorderRadius.circular(5),
                 fieldHeight: 50,
@@ -86,7 +89,7 @@ class VerifyOTPView extends GetView<AuthController> {
                 controller.otp.value = value;
               },
               beforeTextPaste: (text) {
-                controller.otp.value = text;
+                controller.otp.value = text!;
                 //if you return true then it will show the paste confirmation dialog. Otherwise if false, then nothing will happen.
                 //but you can show anything you want here, like your pop up saying wrong paste format or etc
                 return true;
@@ -94,48 +97,57 @@ class VerifyOTPView extends GetView<AuthController> {
             ).pOnly(bottom: 10),
             InkWell(
               onTap: () {
-                if (controller.otp.value.isNotEmpty && controller.otp.value.length == 5) {
+                if (controller.otp.value.isNotEmpty &&
+                    controller.otp.value.length == 5) {
                   controller.verifyOtp();
                 } else {
                   Fluttertoast.showToast(
                     msg: commonService.labelData.value.data.otpValidation,
-                    backgroundColor: Get.theme.errorColor,
+                    backgroundColor: errorColor,
                     gravity: ToastGravity.TOP,
                   );
                 }
               },
               child: Container(
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(10),
-                  gradient: LinearGradient(
-                    colors: [
-                      Get.theme.colorScheme.primary,
-                      Get.theme.colorScheme.secondary,
-                    ],
-                    begin: const FractionalOffset(0.0, 0.0),
-                    end: const FractionalOffset(0.0, 1.0),
-                    stops: const [0.0, 1.0],
-                    tileMode: TileMode.clamp,
+                  padding:
+                      EdgeInsets.only(left: 16, right: 16, top: 12, bottom: 12),
+                  decoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(100),
+                    gradient: LinearGradient(
+                      colors: [
+                        Get.theme.colorScheme.primary,
+                        Get.theme.colorScheme.secondary,
+                      ],
+                      begin: const FractionalOffset(0.0, 0.0),
+                      end: const FractionalOffset(0.0, 1.0),
+                      stops: const [0.0, 1.0],
+                      tileMode: TileMode.clamp,
+                    ),
                   ),
-                ),
-                child: commonService.labelData.value.data.verify.text.uppercase
-                    .textStyle(
-                      Get.textTheme.headline2.copyWith(
-                        color: Get.theme.highlightColor,
-                      ),
-                    )
-                    .make()
-                    .pSymmetric(h: 20, v: 15),
-              ),
+                  child: Text(
+                    commonService.labelData.value.data.verify.toUpperCase(),
+                    style: TextStyle(
+                        color: Get.theme.highlightColor, fontSize: 18),
+                  )),
             ).objectCenterLeft().pOnly(bottom: 20),
             Obx(
               () => !controller.bHideTimer.value
                   ? Row(
                       mainAxisAlignment: MainAxisAlignment.start,
                       children: [
-                        commonService.labelData.value.data.didntRecvOtp.text.center.textStyle(Get.textTheme.bodyText1.copyWith(color: Get.theme.indicatorColor.withOpacity(0.8))).make(),
+                        commonService
+                            .labelData.value.data.didntRecvOtp.text.center
+                            .textStyle(bodyText1.copyWith(
+                                color:
+                                    Get.theme.indicatorColor.withOpacity(0.8)))
+                            .make(),
                         const SizedBox(width: 10),
-                        commonService.labelData.value.data.resendOtp.text.center.textStyle(Get.textTheme.headline3.copyWith(color: Get.theme.indicatorColor.withOpacity(0.8))).make().onInkTap(() {
+                        commonService.labelData.value.data.resendOtp.text.center
+                            .textStyle(headline3().copyWith(
+                                color:
+                                    Get.theme.indicatorColor.withOpacity(0.8)))
+                            .make()
+                            .onInkTap(() {
                           if (!controller.bHideTimer.value) {
                             controller.sendOtpToEmail(false);
                             controller.startTimer();
@@ -148,11 +160,25 @@ class VerifyOTPView extends GetView<AuthController> {
                       children: [
                         SizedBox(
                           width: 75,
-                          child: commonService.labelData.value.data.resendIn.text.textStyle(Get.textTheme.bodyText1.copyWith(fontSize: 16, color: Get.theme.indicatorColor.withOpacity(0.8))).center.make(),
+                          child: commonService
+                              .labelData.value.data.resendIn.text
+                              .textStyle(bodyText1.copyWith(
+                                  fontSize: 16,
+                                  color: Get.theme.indicatorColor
+                                      .withOpacity(0.8)))
+                              .center
+                              .make(),
                         ),
                         SizedBox(
                           width: 50,
-                          child: "${controller.countTimer.value}s".text.textStyle(Get.textTheme.headline1.copyWith(color: Get.theme.indicatorColor.withOpacity(0.8))).center.make().objectCenterRight(),
+                          child: "${controller.countTimer.value}s"
+                              .text
+                              .textStyle(headline1().copyWith(
+                                  color: Get.theme.indicatorColor
+                                      .withOpacity(0.8)))
+                              .center
+                              .make()
+                              .objectCenterRight(),
                         ),
                       ],
                     ),
